Exam Grade Calculator
This program takes the student number and the number of assignments as input from the user and performs calculations based on the following conditions: If the number of assignments is 0, it calculates the weighted average using 60% of the midterm and 40% of the final grades. If there are assignments, it calculates the weighted average using 35% of the midterm, 40% of the final, and 25% of the assignment average.